third_party/WebKit/LayoutTests/platform/win/webaudio/decodeAudioData/decode-audio-data-basic-expected.txt

CONSOLE ERROR: Uncaught (in promise) loadFile: Request failed when loading ../resources/media/invalid-audio-file.txt. ()
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = undefined
PASS # AUDIT TASK RUNNER STARTED. 
PASS > [preload-arraybuffer]  
Harness: the test ran to completion.
